+++ das-client/pom.xml
@@ -19,7 +19,6 @@
 		<commons-lang-version>2.5</commons-lang-version>
 		<easymock-version>2.3</easymock-version>
 		<slf4j-api-version>1.7.12</slf4j-api-version>
-		<cglib-nodep-version>2.1_3</cglib-nodep-version>
 		<libthrift-version>0.12.0</libthrift-version>
 		<guava-version>23.0</guava-version>
 		<spring-version>4.3.8.RELEASE</spring-version>
@@ -55,9 +54,9 @@
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
-			<groupId>cglib</groupId>
-			<artifactId>cglib-nodep</artifactId>
-			<version>${cglib-nodep-version}</version>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-aop</artifactId>
+			<version>${spring-boot-version}</version>
 		</dependency>
 		<dependency>
 			<groupId>commons-codec</groupId>
