--- das-client/src/main/java/com/ppdai/das/client/transaction/TransactionalIntercepted.java
@@ -1,10 +1,0 @@
-package com.ppdai.das.client.transaction;
-
-/**
- * Just a flag to indicate that the annotation is correct intercepted by DalTransactionInterceptor.
- * IMPORTANT: NEVER implement this interface in your own class!!!
- *
- * @author jhhe
- */
-public interface TransactionalIntercepted {
-}
