+++ das-server/pom.xml
@@ -2,13 +2,13 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.ppdai.das</groupId>
     <artifactId>das-server</artifactId>
-    <version>2.1.1-SNAPSHOT</version>
+    <version>2.3.5-SNAPSHOT</version>
     <properties>
         <ver.version>${project.version}</ver.version>
         <file_encoding>UTF-8</file_encoding>
         <junit-version>4.4</junit-version>
         <slf4j-api-version>1.7.12</slf4j-api-version>
-        <das-client-version>2.2.2-SNAPSHOT</das-client-version>
+        <das-client-version>2.3.5-SNAPSHOT</das-client-version>
     </properties>
     <dependencies>
         <dependency>
